+++ lib/cache.php
@@ -7,20 +7,21 @@ class Cache {
 		$this->kirby = $kirby;
 	}
 
-	function get() {
-		return $this->kirby->get('option', 'bpr');
+	function getOption($option) {
+		return $this->kirby->get('option', $option);
 	}
 
-	function getItem($template) {
-		$cache = $this->get();
+	function get($template, $option) {
+		$cache = $this->getOption($option);
 		if(isset($cache[$template])) {
 			return $cache[$template];
 		}
 	}
 
-	function set($template, $data) {
-		$cache = $this->get();
+	function set($template, $data, $option) {
+		$cache = $this->getOption($option);
 		$cache[$template] = $data;
-		$this->kirby->set('option', 'bpr', $cache);
+		$this->kirby->set('option', $option, $cache);
+		return $data;
 	}
 }