+++ lib/definitions.php
@@ -7,11 +7,11 @@ class Definitions {
 	function parse($array) {
 		global $kirby;
 		$this->kirby = $kirby;
-		$array = $this->definitions($array);
+		$array = $this->walkField($array);
 		return $array;
 	}
 
-	function definitions($array) {
+	function walkField($array) {
 		if(array_key_exists('fields', $array)) {
 			foreach($array['fields'] as $key => $field) {
 				if(is_string($field)) {
@@ -20,17 +20,21 @@ class Definitions {
 					if(!empty($field['extends'])) {
 						$array['fields'][$key] = $this->setExtends($field, $key);
 					}
-					// Structure field
-					if(array_key_exists('fields', $field)) {
-						foreach($field['fields'] as $key2 => $item) {
-							if(is_string($item)) {
-								$array['fields'][$key]['fields'][$key2] = $this->setDefinition($item, $key2);
-							} elseif(is_array($item)) {
-								if(!empty($item['extends'])) {
-									$array['fields'][$key]['fields'][$key2] = $this->setExtends($item, $key2);
-								}
-							}
-						}
+					$array = $this->walkStructure($field, $key, $array);
+				}
+			}
+		}
+		return $array;
+	}
+
+	function walkStructure($field, $key, $array) {
+		if(array_key_exists('fields', $field)) {
+			foreach($field['fields'] as $key2 => $item) {
+				if(is_string($item)) {
+					$array['fields'][$key]['fields'][$key2] = $this->setDefinition($item, $key2);
+				} elseif(is_array($item)) {
+					if(!empty($item['extends'])) {
+						$array['fields'][$key]['fields'][$key2] = $this->setExtends($item, $key2);
 					}
 				}
 			}
@@ -39,54 +43,32 @@ class Definitions {
 	}
 
 	function setExtends($part, $key) {
-		$Cache = new Cache();
-		$name = $part['extends'];
-		$cachekey = $name . '.definition';
-		$buffer = $part;
-		unset($buffer['extends']);
-
-		if(!empty($Cache->getItem($cachekey))) {
-			$data = $Cache->getItem($cachekey);
-		} else {
-			$data = $this->find($name);
-			$Cache->set($cachekey, $data);
-		}
-		$part = array_merge($data, $buffer);
-		return $part;
+		$buffer = $part; unset($buffer['extends']);
+		return array_merge($this->cache($part['extends']), $buffer);
 	}
 
 	function setDefinition($part, $key) {
-		$Cache = new Cache();
-		$cachekey = $part . '.definition';
+		return $this->cache($part);
+	}
 
-		if(!empty($Cache->getItem($cachekey))) {
-			$data = $Cache->getItem($cachekey);
+	function cache($name) {
+		$Cache = new Cache();
+		$cache_key = 'blueprint.reader.definitions';
+
+		if($Cache->get($name, $cache_key)) {
+			$data = $Cache->get($name, $cache_key);
+
 		} else {
-			$data = $this->find($part);
-			$Cache->set($cachekey, $data);
+			$data = $this->find($name);
+			$Cache->set($name, $data, $cache_key);
 		}
 		return $data;
 	}
 
 	function find($field) {
-		$match = $this->findByRoot($field);
-		if($match) return $match;
-		return $this->findByRegistry($field);
-	}
-
-	function findByRoot($field) {
-		$filepath = $this->kirby->roots->blueprints . DS . 'fields' . DS . $field . '.yml';
-
-		if(f::exists($filepath)) {
-			return yaml::read($filepath);
-		}
-	}
-
-	function findByRegistry($field) {
-		$blueprints = kirby()->get('blueprint');
-		$definition_key = 'fields/' . $field;
-		if(isset($blueprints[$definition_key]) && f::exists($blueprints[$definition_key])) {
-			return yaml::read($blueprints[$definition_key]);
+		$definition = blueprint('fields/' . $field);
+		if(isset($definition)) {
+			return yaml::read($definition);
 		}
 	}
 }